+++ History.md
@@ -5,6 +5,7 @@
 - Online voting functionality was added. Admins can now define voting blocks, where users can vote on amendments and motions to be adopted or rejected. A documentation about this feature is located at https://sandbox.motion.tools/help#votings .
 - Submitted amendments can now optionally have tags, too, if set up in the motion types.
 - The accessibility was improved in some parts, especially drop-down selections, while reducing the page load time.
+- A new amendment numbering scheme was introduced more appropriate for english environments: "M1 A1" (Motion number + Amendment number)
 - Bugfix: If an amendment was assigned to an agenda item, the agenda item could not be deleted before the amendment was un-assigned first.
 - Bugfix: an empty "Supporting" section was shown in motions and amendments, if only the "liking" function was enabled.
 - Support for Internet Explorer was dropped.